doc/unibi_{get,set}_ext_*_name: Fix man's “can't break line” warnings
As shown by
“env MANWIDTH=80 man -l <(pod2man doc/unibi_set_ext_num_name.pod) >/dev/null”
man complains about not being able to break a line in the converted man
pages.  This is due to the concatenation of multiple function names
together without a wrap point.  Changing the “/” to a comma separated
list, as already done in the previous paragraph, fixes the warning.

Signed-off-by: James McCoy <jamessan@debian.org>